@@ -2,24 +2,17 @@ from django.conf.urls import url
 
 urlpatterns = []
 
-
-from authentication_app.views import GenerateOTPForSignupView
-from authentication_app.views import VerifyOTPForSignupView
+from authentication_app.views import GenerateOTPView
 
 urlpatterns += [
-
-    url(r'^generate-otp-for-signup', GenerateOTPForSignupView.as_view()),
-    url(r'^verify-otp-for-signup', VerifyOTPForSignupView.as_view())
-
+    url(r'^generate-otp', GenerateOTPView.as_view()),
 ]
 
-
-from authentication_app.views import GenerateOTPView
-from authentication_app.views import VerifyOTPView
+from authentication_app.views import VerifyOTPAndChangePasswordView,\
+    VerifyOTPAndCreateSchool, VerifyOTPAndSignupView
 
 urlpatterns += [
-
-    url(r'^generate-otp', GenerateOTPView.as_view()),
-    url(r'^verify-otp', VerifyOTPView.as_view())
-
+    url(r'^verify-otp-and-signup', VerifyOTPAndSignupView.as_view()),
+    url(r'^verify-otp-and-change-password', VerifyOTPAndChangePasswordView.as_view()),
+    url(r'^verify-otp-and-create-school', VerifyOTPAndCreateSchool.as_view()),
 ]
