@@ -43,11 +43,9 @@ export class SidebarComponent implements OnInit {
     session_list = [];
 
     constructor(private router: Router, private notificationService: NotificationService, private schoolService: SchoolService) {
+        // We are using this routeReuseStrategy because the DashBoard Component should not re-render when changing pages.
         this.router.routeReuseStrategy.shouldReuseRoute = function (future: any, curr: any) {
-            if (curr._routerState.url.includes('print:print') || future._routerState.url.includes('print:print')) {
-                return curr.routeConfig === future.routeConfig;
-            }
-            return false;
+            return (curr.routeConfig === future.routeConfig) || future.data.reuse;
         };
     }
 
@@ -80,9 +78,15 @@ export class SidebarComponent implements OnInit {
             this.session_list = value;
         });
         EmitterService.get('initialize-router').subscribe((value) => {
-            this.router.navigateByUrl(
-                this.router.createUrlTree([this.user.section.route + '/' + this.user.section.subRoute], { queryParams: value.queryParams })
-            );
+            // Navigating To '/' before any other route - because :
+            // We have used routeReuseStrategy so if the url is same the page won't reload,
+            // To overcome that case we are navigating to '/' first and then the corresponding route.
+            this.router.navigateByUrl('/', {skipLocationChange: true}).then(() => {
+                this.router.navigateByUrl(
+                    this.router.createUrlTree([Constants.dashBoardRoute + '/' + this.user.section.route + '/' + this.user.section.subRoute],
+                        {queryParams: value.queryParams})
+                );
+            });
         });
     }
 
@@ -139,5 +143,6 @@ export class SidebarComponent implements OnInit {
         this.user.isAuthenticated = false;
         this.user.jwt = '';
         this.user.emptyUserDetails();
+        this.router.navigate(['login']);
     }
 }
