@@ -7,7 +7,7 @@ frontend/test_mobile_frontend.sh    //start frontend for mobile testing
 
 export const environment = {
     production: false,
-    versionCheckURL: 'www.korangle.com/version.json',
+    versionCheckURL: 'app.korangle.com/version.json',
     DJANGO_SERVER: 'http://' + window.location.hostname + ':8000', //for mobile + normal debugging over LAN
     CASHFREE_CHECKOUT_URL: 'https://test.cashfree.com/billpay/checkout/post/submit',
     api_version: '/v2.0',
