@@ -2,8 +2,8 @@
 
 export const environment = {
     production: true,
-    versionCheckURL: 'www.korangle.com/version.json',
-    DJANGO_SERVER: 'https://www.korangle.com:8443',
+    versionCheckURL: 'app.korangle.com/version.json',
+    DJANGO_SERVER: 'https://app.korangle.com:8443',
     CASHFREE_CHECKOUT_URL: 'https://www.cashfree.com/checkout/post/submit',
     api_version: '/v2.0',
 };
