@@ -1,6 +1,6 @@
 export const environment = {
     production: true,
-    versionCheckURL: 'www.korangle.com/version.json',
+    versionCheckURL: 'app.korangle.com/version.json',
     DJANGO_SERVER: 'https://' + window.location.hostname + ':8443',
     CASHFREE_CHECKOUT_URL: 'https://test.cashfree.com/billpay/checkout/post/submit',
     api_version: '/v2.0',
