@@ -2,7 +2,7 @@
 
 export const environment = {
     production: false,
-    versionCheckURL: 'www.korangle.com/version.json',
+    versionCheckURL: 'app.korangle.com/version.json',
     DJANGO_SERVER: 'http://localhost:8000', // normal debugging only (localhost)
     CASHFREE_CHECKOUT_URL: 'https://test.cashfree.com/billpay/checkout/post/submit',
     api_version: '/v2.0',
