@@ -4,6 +4,6 @@ export const ID_CARD_LAYOUT_BASE_MODEL = {
     parentSchool: 0,
     name: '-----',
     content: '[]',
-    backgroundImage: 'https://www.korangle.com/assets/img/plain-id-card.jpg',
+    backgroundImage: 'https://app.korangle.com/assets/img/plain-id-card.jpg',
 };
 
